Improvements to AiPathfinder, Additional enemies, Maps have more SpawnPoints now, Movement strings changed to enum, some directory repositions, BackgoundAsset is now LevelAsset, same with BackgroundBuilder

+++ LeGame/Models/Characters/Character.cs
@@ -10,7 +10,7 @@
 
     public abstract class Character : GameObject, ICharacter, ICollidable
     {
-        protected Character(Vector2 position,string type, int maxHealth, int currentHealth, int speed, int hitCooldown, ILevel level)
+        protected Character(Vector2 position,string type, int maxHealth, int currentHealth, float speed, int hitCooldown, ILevel level)
             : base(position, type)
         {
             this.MaxHealth = maxHealth;
@@ -37,7 +37,7 @@
 
         public int CurrentHealth { get; set; }
 
-        public int Speed { get; set; }
+        public float Speed { get; set; }
 
         public float FacingAngle { get; set; }
